More updates
- lots of little fixes

Moved the timer, cars and checkpoints into Track. Added init methods and such.

Altered the physics of the car to work with Peter's methods. We can now give car an angle to drive at, and it will accelerate at that angle. I reset the velocity upon turning to stop the cool 'drifting' effect

Added a button to randomly turn the car as a demo